Fix setting foreground colour when using pos-tip-show
pos-tip-show would only pass the tip-color to pos-tip-show-no-propertize,
which resulted in applying it only to frame properties by x-show-tip, but not
to the font of the tooltip itself. This seemed to work actually, but only as
long as user did not choose a custom colour theme.

Otherwise it was impossible to set text foreground colour using
pos-tip-show. (You could not just propertize the text using face as
pos-tip-show replaces the face property of the input text.)

The solution is based on how tooltip.el uses x-show-tooltip: it does set
foreground-color and background-color frame parameters, but it also
additionally applies the tooltip face to the tooltip text.